+++ noodles-bcf/src/async/reader.rs
@@ -319,7 +319,7 @@ where
     where
         I: BinningIndex,
     {
-        use crate::reader::resolve_region;
+        use crate::io::reader::resolve_region;
 
         let reference_sequence_id = resolve_region(contig_string_map, region)?;
         let chunks = index.query(reference_sequence_id, region.interval())?;