+++ noodles-bcf/src/lib.rs
@@ -6,14 +6,12 @@
 mod r#async;
 
 pub mod header;
-pub mod indexed_reader;
 pub mod io;
 pub mod lazy;
-pub mod reader;
 pub(crate) mod record;
 pub mod writer;
 
-pub use self::{indexed_reader::IndexedReader, reader::Reader, writer::Writer};
+pub use self::writer::Writer;
 
 #[cfg(feature = "async")]
 pub use self::r#async::Reader as AsyncReader;