+++ noodles-bcf/examples/bcf_count_async.rs
@@ -15,7 +15,7 @@ async fn main() -> io::Result<()> {
     let mut reader = File::open(src).await.map(bcf::r#async::io::Reader::new)?;
     reader.read_header().await?;
 
-    let mut records = reader.lazy_records();
+    let mut records = reader.records();
     let mut n = 0;
 
     while records.try_next().await?.is_some() {